fix: check read() read entire buffer size
The following fix is to fix a security issue in which if read() is called and a process modifies the file at the same time to finish early, there will be end bytes of the buffer that read() did not copy to or 0 out, thus disclosing process memory. Some systems do 0 out on malloc() nowadays, but this is not all systems (and is not a true fix for this) so I am PRing this patch.

Apologies for not reporting privately, but at the time of PR there is no private security reporting enabled.