feat: add sensitive pointer guarding
This is a new security feature that while does not fix any vulnerabilities is meant to serve as a mitigation to restrict existing vulnerabilities. If someone ever gets arbitrary write access to __DATA segment, this means that they can simply overwrite flags for other security measures to bypass all security. The following commit introduces pointer guarding, which guards the sensitive flags to be read-only whenever possible.