Memory leak fix? (to be confirmed)
This fix ensures the operation is only tied to the current scope. And it shows that it works. We have on one simple app compilation in the end 90mb instead of 200mb of memory usage. I suspect the rest is caused by LD which is GNUs linker. I hope to the release I will already have a own managed linker implemented.